Refactors ViewerPaginator
Refactor paginator state
