+++ src/lib/stores/viewer.ts
@@ -1,0 +1,5 @@
+import { writable, type Writable } from "svelte/store";
+import type { Note, ViewerMode } from "../api/types";
+
+export const currentPage: Writable<number> = writable(1);
+export const activeNote: Writable<Note> = writable(null);
